requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
cryptography==42.0.5
SQLAlchemy==2.0.23